Updating data types for models

+++ kryptokrona-node/src/main/kotlin/org/kryptokrona/sdk/node/model/request/block/BlocksListRequest.kt
@@ -42,6 +42,6 @@ import kotlinx.serialization.Serializable
 data class BlocksListRequest(
     val jsonrpc: String = "2.0",
     val method: String,
-    val params: Map<String, Int>
+    val params: Map<String, Long>
 )
