+++ kryptokrona-node/src/main/kotlin/org/kryptokrona/sdk/node/model/request/block/GetBlockHashRequest.kt
@@ -42,5 +42,5 @@ import kotlinx.serialization.Serializable
 data class GetBlockHashRequest(
     val jsonrpc: String = "2.0",
     val method: String,
-    val params: List<Int>
+    val params: List<Long>
 )
